envsetup.sh: Set CCACHE_DIR from cache.txt if present
Always check for a root-level cache.txt file and set CCACHE_DIR from
its contents if found. If the file does not exist or is empty, fall
back to $HOME/.ccache. This ensures that custom ccache locations are
respected even if CCACHE_EXEC is already defined.

Also ensures that CCACHE_DIR is created if missing.

Change-Id: Ibd0c4642daa995fffb1677e283b43a1bb4a0c8bf
Signed-off-by: klozz <carlosj@klozz.dev>